requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
deprecation==2.1.0
distlib==0.3.9
Django==5.2
django-allauth==65.7.0
django-classy-tags==4.1.0
django-cms==4.1.6
django-cors-headers==4.0.0
django-formtools==2.5.1
django-grappelli==4.0.1
django-sekizai==4.1.0
django-treebeard==4.7.1
djangocms-admin-style==3.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.16.1
frozenlist==1.6.0
gotrue==2.12.0
gunicorn==23.0.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.5
idna==3.10
iniconfig==2.1.0
multidict==6.4.3
nodeenv==1.9.1
oauthlib==3.2.2
packaging==24.2
pillow==11.2.1
platformdirs==4.3.6
pluggy==1.5.0
postgrest==1.0.1
pre_commit==4.0.1
propcache==0.3.1
psycopg==3.1.9
psycopg2-binary==2.9.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.0
virtualenv==20.29.1
websockets==14.2
yarl==1.20.0
